refactor: move recording worklet to @waveform-playlist/worklets package
- Copy recording-processor.worklet.ts and its tests to worklets package
- Update useRecording.ts to import recordingProcessorUrl from worklets
- Simplify recording tsup config to single entry (no worklet build)
- Add @waveform-playlist/worklets as dependency of recording package
- Remove old worklet directory from recording package
