+++ packages/recording/package.json
@@ -52,7 +52,8 @@
   "dependencies": {
     "@waveform-playlist/core": "workspace:*",
     "@waveform-playlist/playout": "workspace:*",
-    "@waveform-playlist/ui-components": "workspace:*"
+    "@waveform-playlist/ui-components": "workspace:*",
+    "@waveform-playlist/worklets": "workspace:*"
   },
   "peerDependencies": {
     "react": "^18.0.0",
