+++ packages/recording/tsup.config.ts
@@ -1,27 +1,13 @@
 import { defineConfig } from 'tsup';
 
-export default defineConfig([
-  // Main package
-  {
-    entry: {
-      index: 'src/index.ts',
-    },
-    format: ['cjs', 'esm'],
-    dts: true,
-    splitting: false,
-    sourcemap: true,
-    clean: true,
-    external: ['tone', 'react', 'react-dom', 'react/jsx-runtime', 'react/jsx-dev-runtime', 'styled-components'],
+export default defineConfig({
+  entry: {
+    index: 'src/index.ts',
   },
-  // AudioWorklet processor (no DTS generation)
-  {
-    entry: {
-      'worklet/recording-processor.worklet': 'src/worklet/recording-processor.worklet.ts',
-    },
-    format: ['cjs', 'esm'],
-    dts: false,
-    splitting: false,
-    sourcemap: true,
-    clean: false,
-  },
-]);
+  format: ['cjs', 'esm'],
+  dts: true,
+  splitting: false,
+  sourcemap: true,
+  clean: true,
+  external: ['tone', 'react', 'react-dom', 'react/jsx-runtime', 'react/jsx-dev-runtime', 'styled-components'],
+});
