+++ packages/worklets/src/index.ts
@@ -1,7 +1,5 @@
-export const meterProcessorUrl = new URL(
-  './worklet/meter-processor.worklet.js',
-  import.meta.url
-).href;
+export const meterProcessorUrl = new URL('./worklet/meter-processor.worklet.js', import.meta.url)
+  .href;
 
 export const recordingProcessorUrl = new URL(
   './worklet/recording-processor.worklet.js',
